R5: Split finance.js into 7 sub-route files
Split the monolithic backend/src/routes/finance.js (1,779 lines) into
focused sub-route files under backend/src/routes/finance/:

- accounts.js — accounts CRUD, group B/F setting, payment method defaults
- categories.js — category CRUD
- transactions.js — list, single, create, bulk-pending, update, delete, refund
- transfers.js — transfer money CRUD
- reconciliation.js — account reconciliation
- statements.js — financial statement + groups statement (JSON + Excel)
- batches.js — credit batch CRUD + transaction management
- helpers.js — shared computeYearBounds utility
- index.js — parent router mounting all sub-routers

All 311 backend tests pass without changes.
